monitor redis for changes
recent_beacons is not used in this app,
I just included it here for convenience so
we can add beacons via rails console for testing.

beacon_monitor will subscribe to a redis channel
and emit SSE events when beacons are added, updated
or removed.

right now we're just polling and emitting the whole list
in a single event for two reasons:

1) the publishing app is not pub/sub in redis, so we have no
channel to subscribe to.

2) i'm not sure yet how we'll identify a beacon signal:
probably beacon-uuid + sensor-id, but need to confirm.
(plus, currently sensor-id is not provided).